refactor edit cavc remand form
